Extract MemoryBudget and MemoryTicket into separate module
Moves the memory budget implementation from prefetch.rs into
chunks/memory_budget.rs for better separation of concerns.
